+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add required field asterisks & validation action to Order field mapping profile (UIDATIMP-1265)
 * Add required info icons to Order field mapping profile (UIDATIMP-1266)
 * Data Import field mapping profile: Enable FOLIO record type = Orders (UIDATIMP-1267)
+* Add required info icons to Holdings field mapping profile (UIDATIMP-1292)
 * Add required info icons to Item field mapping profile (UIDATIMP-1293)
 * Allow edit and duplicate Default - Create SRS MARC Authority job profile (UIDATIMP-1318)